requirement.txt
fastapi
uvicorn[standard]
PyJWT
passlib